Resolve Windows x64 issues.

+++ lib-src/portmixer/src/px_win_ds.c
@@ -96,8 +96,8 @@ int OpenMixer_Win_DirectSound(px_mixer *Px, int index)
    ULONG bytes;
    LPGUID guidIn;
    LPGUID guidOut;
-   UINT deviceIn = -1;
-   UINT deviceOut = -1;
+   ULONG deviceIn = -1;
+   ULONG deviceOut = -1;
    int ret = FALSE;
 
    guidIn = PaWinDS_GetStreamInputGUID(Px->pa_stream);
